Require Node 10 and adopt async-await
The puppeteer API is optimised for Node 8+ and makes heavy use
of async functions. These will be much easier and cleaner to use
on newer Node versions.

I've just finished creating Debian Stretch images with Node 10
for CI, which we can use going forward. Given this is mainly
developer oriented and will only be deployed on newer hardware
within WMF CI that supports Node 10 already, no need to support
ES5 or Node 6 (which will be EOL by the end of the year anyway).

Bug: T133646
Change-Id: Id55c84fe2a975cddffcbd0189ba32adf980392b6