build: Update dependencies (depth: 2)
Follows-up 235abecaa09, which did not completely solve the npm-audit
warning about vulnerabilities in 'handlebars'. It fixed the dep
from nyc, but not the dep from jsdoc-to-markdown.

This is because it is an indirect dependency. We are already
using the latest version of jsdoc-to-markdown, and it already
allows use of newer handlebars, we just need to tell npm to
re-expand the flexible version ranges there as well.

This is result from 'npm update --depth 2'.

> jsdoc-to-markdown > dmd > handlebars

Change-Id: I0ea702dcbddcfdace1269203c0c5f6d177ac2a3a