conductor: Support plain objects in config.scenarios
Previously, the list of scenarios had to be an array, which has
only numerical keys (implied).

This meant that the build output wasn't very descriptive.
E.g. it would say "scenario 0" or "scenario 1". It is hard to
remember which one is which (e.g. main page, edit form etc.)

The comparison JSON format already used plain objects.

The record and config JSON formats, now use plain objects
as well.

Array is still supported as well, will be treated the same way
as before ("0", "1" etc.). But users are recommended to use
string keys instead, which will be more descriptive and useful.

Bug: T216854
Change-Id: Ied19f92a64f3c4b90d0c5c808236f5c7c7965511