test: Simplify unit testing of config
Turn this from a hacky integration test into a unit test.

It was only testing the call in conductor.record() to is.config(),
which we can instead test directly.

Previously it was making sure the rest of the function did not
run by making puppeteer.launch() fail, which led to a "valid"
configuration being asserted by a failing command (reaching the
stubbed out puppeteer call). This is all gone now.

Change-Id: I81e98aa85d17232941b326b4d769bfff3b3b88af