Add Mann-Whitney U test to compute module
Add Mann-Whitney U test for estimating whether a commit worsens
performance, as an alternative to checking whether the standard
deviations overlap.

This only adds the new test and does not use it for reporting yet.
Using the new test for reports will require making some judgements
such as whether to remove the old test completely or keep it for
reporting basic stats.

Bug: T223977
Change-Id: Ie80631af582af9dc1c8e22722df481764cf5b780