reports: Raise navtiming and paint thesholds to +100ms and +10ms
These are currently quite variable and cause of the most false-negatives.
For the time being, just increase these by 10x.

This doesn't change the contents of the summary report presented
in the Jenkins job. It just means that we only raise the red
flag when the transfer sizes regress (over which we have high
confidence).

The timing-based metrics affected by this issue (navtiming loadEventEnd,
and paint firstPaint) are as variable as they are because they
are a high-level span over many individual smaller processes.

Once we have metrics for the individual smaller chunks and steps
involved with loading a page, we can have lower thresholds again
as they are less prone to long stalls on CPU/IO time. For example,
the 'onLoad' callback is a very specific callback start-to-end and
has almost no variance. Similarly, the mw.loader.profiling data
is also measured very locally which is always processed at once
(which means we start and end the timer within a single processing
chunk).

Bug: T227004
Change-Id: I5fc7ebda6fc944697f04b82b2f11a5362a537f0b