reports: Raise threshold for bytes transfers from 1B to 3B
The +1B or +2B triggers falsely sometimes due to standard variance
in headers (I suspect the "Date" header), the body sizes are constant
but the other data varies.

This should make it slightly less sensitive whilst still being close
to zero-tolerant, given that Fresnel is only meant to give a hint
that somethig might be increasing cost unexpectedly, it's not blocking
but should always flag an increase in cost, no matter how small.

Raise the total from 1 to 5B, so that we allow +2 on both JS and CSS.

Change-Id: If010f885effc8e3bbd3646c28f96681193784977