compute: Rename compareStdev() to diffStdev()
This makes the return value less confusing, 0 means there is no
difference and positive value means there is an upward difference
detected. This is returning the result of expecting a difference,
not of comparing how similar they are.

E.g. compare-like functions tend to return 1 if they are the same
and 0 if they were different, whereas this is the opposite.

This is in preparation for adding a diffMannWhitney() function
that will inverse mannWhitney() to behave the same way.

Bug: T223977
Change-Id: I269a7023a89716b238355bdc9d3bd34cf627cd82