compute: Add diffMannWhitney()
Before can use mannWhitney() as replacement of Stddev for deciding
whether to emit a perf warning in the Fresnel reports, we need
to invert this value.

The compare+threshold abstraction requires the threshold to be
the highest numerical value to tolerate, which, if exceeded
results in a perf warning. E.g. if difference is more than 10ms
or more than 50 bytes.

To use a probability value, we need the probability to get higher
when the results are more confident, but right now the probability
is not for the values being higher, but for them being the same.
Invert this so that we can e.g. set a threshold of 0.90 and then
it will warn if the probability is between 0.90 and 1.0 that the
values got higher.

Bug: T223977
Change-Id: Ic6b39c189756e86589483acb897adc8852110bf7