build: Update eslint-config-wikimedia to 0.20.0
== Non-obvious fixes ==

no-use-before-define:
* (src/compute.js) 'ranks' was used before it was defined

node/no-deprecated-api:
* (src/reports/transfer.js) 'url.parse' was deprecated. Use 'url.URL' instead.

== Disabled ==

jsdoc/no-undefined-types:
* (src/recorder.js) The type 'Probe' is undefined
* (src/recorder.js) The type 'Writer' is undefined

This isn't smart enough yet it seems.

jsdoc/valid-types:
* Syntax error in namepath: puppeteer/Browser
* Syntax error in type: external:puppeteer/Browser

Looks like it doesn't understand understands. I figured maybe its wants
me to use a dot instead of a slash here, but while that satisfies
eslint-plugin-jsdoc, it breaks jsdoc. The slash is required here for
it to have the effect of treating it as an external type where the
link goes to the specified URL.

This repo also runs actual jsdoc in the test command, which still
authoritatively validate these without the lint rule.

Change-Id: I83b16d2b1cec098398e068bd405f5e24fd4bcef4