build: Use lockVersion 2 to fix puppeteer postinstall
Installing fresnel package continues to be broken on npm 7+ even
on latest Node with all default settings. And indeed simply running
`npm install fresnel@1.1.0` on latest nodejs/node with empty PWD
and cache, is still broken. (That is, puppeteer is fetched without
triggering the postinstall script that downloads Chromium, and yes
this isn't about whether the postinstall script has its output shown,
that is a separate and intentional change in npm 7, the script
genuinely doesn't run).

This affects all npm packages that have a lock file with lockVersion 1,
and a dependency that needs a postinstall script, such as Puppeteer.

== Problem

One of the lockVersion 2 features is that npm now caches inside the
lock file whether it had to run a postinstall script (instead of
checking each of the dependencies' package.json files), in the form
of `"hasInstallScript": true`.

But, it seems that its compatibility layer for lockVersion 1 just
defaults to pretending no packages have install scripts and so
doesn't run them.

Upstream issues, as of writing most still open and unresolved:

* https://github.com/npm/cli/issues/1905
* https://github.com/npm/cli/issues/2853
* https://github.com/npm/cli/issues/3123
* https://github.com/npm/cli/issues/3807
* https://github.com/puppeteer/puppeteer/issues/6560
* https://github.com/puppeteer/puppeteer/issues/6608

== lockVersion 1

The reason fresnel is published with lockVersion 1, is because I ran
"npm install" during the release commit from fresh-node10 which has
npm 6, where the latest lockVersion was lockVersion 1.

The reason I used fresh-node10 is that the latest Quibble still comes
with Node.js 10, so it seemed best to install it that way in case it
would otherwise break something.

Given the above, and given that Quibble already accidentally changed
its npm version from 6 to 7 (despite still being on Node.js 10), it
should be safe to publish this with lockVersion 2.

To generate this, I ran `npm install` from fresh-node (Node 12/npm 7),
which I expect will work in Quibble which comes with Node 10/npm 7.

Bug: T293150
Change-Id: I00556a3d63651219d8f041fe666db3844d8729e4